requirement.txt
google-generativeai
python-dotenv
Pillow
PyYAML